+++ app/src/main/res/layout/lista_completa_pessoa_fisica.xml
@@ -8,89 +8,119 @@
         android:id="@+id/txtViewNome"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="Nome"/>
+        android:text="Nome"
+        android:textSize="22sp"
+        android:layout_marginTop="7dp"/>
 
     <TextView
         android:id="@+id/txtViewCpf"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="CPF"/>
+        android:text="CPF"
+        android:textSize="22sp"
+        android:layout_marginTop="7dp"/>
 
     <TextView
         android:id="@+id/txtViewRg"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="RG"/>
+        android:text="RG"
+        android:textSize="22sp"
+        android:layout_marginTop="7dp"/>
 
     <TextView
         android:id="@+id/txtViewCnh"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="CNH"/>
+        android:text="CNH"
+        android:textSize="22sp"
+        android:layout_marginTop="7dp"/>
 
     <TextView
         android:id="@+id/txtViewSexo"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="Sexo"/>
+        android:text="Sexo"
+        android:textSize="22sp"
+        android:layout_marginTop="7dp"/>
 
     <TextView
         android:id="@+id/txtViewDataNasc"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="Data Nascimento"/>
+        android:text="Data Nascimento"
+        android:textSize="22sp"
+        android:layout_marginTop="7dp"/>
 
     <TextView
         android:id="@+id/txtViewTelefone"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="Telefone"/>
+        android:text="Telefone"
+        android:textSize="22sp"
+        android:layout_marginTop="7dp"/>
 
     <TextView
         android:id="@+id/txtViewEndereco"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="Endereco"/>
+        android:text="Endereco"
+        android:textSize="22sp"
+        android:layout_marginTop="7dp"/>
 
     <TextView
         android:id="@+id/txtViewNumero"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="Numero"/>
+        android:text="Numero"
+        android:textSize="22sp"
+        android:layout_marginTop="7dp"/>
 
     <TextView
         android:id="@+id/txtViewCidade"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="Cidade"/>
+        android:text="Cidade"
+        android:textSize="22sp"
+        android:layout_marginTop="7dp"/>
 
     <TextView
         android:id="@+id/txtViewEstado"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="Estado"/>
+        android:text="Estado"
+        android:textSize="22sp"
+        android:layout_marginTop="7dp"/>
 
     <TextView
         android:id="@+id/txtViewBairro"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="Bairro"/>
+        android:text="Bairro"
+        android:textSize="22sp"
+        android:layout_marginTop="7dp"/>
 
     <TextView
         android:id="@+id/txtViewCep"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="CEP"/>
+        android:text="CEP"
+        android:textSize="22sp"
+        android:layout_marginTop="7dp"/>
 
     <TextView
         android:id="@+id/txtViewEmail"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="Email"/>
+        android:text="Email"
+        android:textSize="22sp"
+        android:layout_marginTop="7dp"/>
 
     <TextView
         android:id="@+id/txtViewProfissao"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="Profissao"/>
+        android:text="Profissao"
+        android:textSize="22sp"
+        android:layout_marginTop="7dp"/>
 </LinearLayout>
